Flowers/requirements.txt
django==5.0.6
psycopg2==2.9.9
pytz==2024.1
google-auth==2.29.0
google-api-python-client==2.129.0